Update to p-value method; new examples
The false-positive (FP) count cutoff method was replaced with simply reporting the binomial p-value, but allowing the user to specify a cutoff using the --p_cutoff argument; this obviates the need for --seq_len and --num_samples arguments. P-values are now included in both the INFO and SAMPLE columns. A minimal example was added. Documentation was updated.